requirements.txt
apscheduler
cryptography
django
django-cors-headers
django-apscheduler
furl
jinja2
scrapy
scrapy-redis
scrapy-splash
python-scrapyd-api
redis
requests
pymongo
pymysql
pyquery
beautifulsoup4
djangorestframework
websocket
pyppeteer

# heroku
dj_database_url
gunicorn
psycopg2-binary
httpie
